ml/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
psycopg2-binary==2.9.7
pickle-mixin==1.0.2